requirements.txt
pandas==1.5.3
scikit-learn==1.2.1
PyYAML==6.0
pandas==1.5.3
dill==0.3.5.1
fastapi==0.94.0
uvicorn==0.21.0
python-multipart==0.0.6
joblib==1.2.0
Flask==2.3.2
cassandra-driver==3.27.0
pytest-shutil==1.7.0
-e .
